// packages/node-sdk/src/model/createEnvironment200Response.ts
/**
 * Ctrlplane API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { CreateEnvironment200ResponseEnvironment } from "./createEnvironment200ResponseEnvironment";
import { RequestFile } from "./models";

export class CreateEnvironment200Response {
  "environment"?: CreateEnvironment200ResponseEnvironment;

  static discriminator: string | undefined = undefined;

  static attributeTypeMap: Array<{
    name: string;
    baseName: string;
    type: string;
  }> = [
    {
      name: "environment",
      baseName: "environment",
      type: "CreateEnvironment200ResponseEnvironment",
    },
  ];

  static getAttributeTypeMap() {
    return CreateEnvironment200Response.attributeTypeMap;
  }
}