// packages/node-sdk/src/model/createReleaseRequest.ts
/**
 * Ctrlplane API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from "./models";

export class CreateReleaseRequest {
  "version": string;
  "deploymentId": string;
  "metadata"?: object;

  static discriminator: string | undefined = undefined;

  static attributeTypeMap: Array<{
    name: string;
    baseName: string;
    type: string;
  }> = [
    {
      name: "version",
      baseName: "version",
      type: "string",
    },
    {
      name: "deploymentId",
      baseName: "deploymentId",
      type: "string",
    },
    {
      name: "metadata",
      baseName: "metadata",
      type: "object",
    },
  ];

  static getAttributeTypeMap() {
    return CreateReleaseRequest.attributeTypeMap;
  }
}