// packages/node-sdk/src/model/deleteTarget200Response.ts
/**
 * Ctrlplane API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from "./models";

export class DeleteTarget200Response {
  "success": boolean;

  static discriminator: string | undefined = undefined;

  static attributeTypeMap: Array<{
    name: string;
    baseName: string;
    type: string;
  }> = [
    {
      name: "success",
      baseName: "success",
      type: "boolean",
    },
  ];

  static getAttributeTypeMap() {
    return DeleteTarget200Response.attributeTypeMap;
  }
}