// packages/node-sdk/src/model/getAgentRunningJob200ResponseInner.ts
/**
 * Ctrlplane API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from "./models";

export class GetAgentRunningJob200ResponseInner {
  "id": string;
  "status": string;
  "message": string;
  "jobAgentId": string;
  "jobAgentConfig": object;
  "externalId": string | null;
  "release"?: object;
  "deployment"?: object;
  "config": object;
  "runbook"?: object;
  "target"?: object;
  "environment"?: object;

  static discriminator: string | undefined = undefined;

  static attributeTypeMap: Array<{
    name: string;
    baseName: string;
    type: string;
  }> = [
    {
      name: "id",
      baseName: "id",
      type: "string",
    },
    {
      name: "status",
      baseName: "status",
      type: "string",
    },
    {
      name: "message",
      baseName: "message",
      type: "string",
    },
    {
      name: "jobAgentId",
      baseName: "jobAgentId",
      type: "string",
    },
    {
      name: "jobAgentConfig",
      baseName: "jobAgentConfig",
      type: "object",
    },
    {
      name: "externalId",
      baseName: "externalId",
      type: "string",
    },
    {
      name: "release",
      baseName: "release",
      type: "object",
    },
    {
      name: "deployment",
      baseName: "deployment",
      type: "object",
    },
    {
      name: "config",
      baseName: "config",
      type: "object",
    },
    {
      name: "runbook",
      baseName: "runbook",
      type: "object",
    },
    {
      name: "target",
      baseName: "target",
      type: "object",
    },
    {
      name: "environment",
      baseName: "environment",
      type: "object",
    },
  ];

  static getAttributeTypeMap() {
    return GetAgentRunningJob200ResponseInner.attributeTypeMap;
  }
}