// packages/node-sdk/src/model/getJob200ResponseApproval.ts
/**
 * Ctrlplane API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { GetJob200ResponseApprovalApprover } from "./getJob200ResponseApprovalApprover";
import { RequestFile } from "./models";

export class GetJob200ResponseApproval {
  "id": string;
  "status": GetJob200ResponseApproval.StatusEnum;
  "approver"?: GetJob200ResponseApprovalApprover | null;

  static discriminator: string | undefined = undefined;

  static attributeTypeMap: Array<{
    name: string;
    baseName: string;
    type: string;
  }> = [
    {
      name: "id",
      baseName: "id",
      type: "string",
    },
    {
      name: "status",
      baseName: "status",
      type: "GetJob200ResponseApproval.StatusEnum",
    },
    {
      name: "approver",
      baseName: "approver",
      type: "GetJob200ResponseApprovalApprover",
    },
  ];

  static getAttributeTypeMap() {
    return GetJob200ResponseApproval.attributeTypeMap;
  }
}

export namespace GetJob200ResponseApproval {
  export enum StatusEnum {
    Pending = <any>"pending",
    Approved = <any>"approved",
    Rejected = <any>"rejected",
  }
}