// packages/node-sdk/src/model/getJob200ResponseApprovalApprover.ts
/**
 * Ctrlplane API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from "./models";

/**
 * Null when status is pending, contains approver details when approved or rejected
 */
export class GetJob200ResponseApprovalApprover {
  "id": string;
  "name": string;

  static discriminator: string | undefined = undefined;

  static attributeTypeMap: Array<{
    name: string;
    baseName: string;
    type: string;
  }> = [
    {
      name: "id",
      baseName: "id",
      type: "string",
    },
    {
      name: "name",
      baseName: "name",
      type: "string",
    },
  ];

  static getAttributeTypeMap() {
    return GetJob200ResponseApprovalApprover.attributeTypeMap;
  }
}